workflows/pyproject.toml
[project]
name = "foundry-hosted-langchain-workflow-demos"
version = "0.1.0"
description = "LangGraph workflow demos for Microsoft Foundry"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-agentserver-langgraph==1.0.0b14",
    "azure-identity>=1.25.3",
    "azure-monitor-opentelemetry>=1.8.4",
    "httpx>=0.28.1",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "python-dotenv==1.1.1",
    "starlette<1.0.0",  # azure-ai-agentserver uses deprecated on_event removed in 1.0
]

[tool.uv]
prerelease = "allow"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
